requirements.txt
scikit-learn
scikit-surprise
numpy<2.0.0